+++ python/lsst/ts/ofc/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # This file is part of ts_ofc.
 #
-# Developed for the LSST Telescope and Site Systems.
+# Developed for Vera Rubin Observatory.
 # This product includes software developed by the LSST Project
 # (https://www.lsst.org).
 # See the COPYRIGHT file at the top-level directory of this distribution