+++ python/lsst/ts/ofc/utils/__init__.py
@@ -22,4 +22,3 @@
 
 from .enums import *
 from .utils import *
-from .rot_1d_array import *