+++ tests/test_ofc.py
@@ -1,6 +1,6 @@
 # This file is part of ts_ofc.
 #
-# Developed for the LSST Telescope and Site Systems.
+# Developed for Vera Rubin Observatory.
 # This product includes software developed by the LSST Project
 # (https://www.lsst.org).
 # See the COPYRIGHT file at the top-level directory of this distribution
@@ -26,7 +26,7 @@ from lsst.ts.ofc import OFC, OFCData, Correction
 from lsst.ts.ofc.utils import get_pkg_root, CorrectionType
 
 
-class TestOFCCalculation(unittest.TestCase):
+class TestOFC(unittest.TestCase):
     """Test the OFCCalculation class."""
 
     def setUp(self):